+++ fidget/src/jit/x86_64/interval.rs
@@ -41,7 +41,7 @@ impl Assembler for IntervalAssembler {
         Self(out)
     }
     fn build_load(&mut self, dst_reg: u8, src_mem: u32) {
-        assert!(dst_reg < REGISTER_LIMIT);
+        assert!((dst_reg as usize) < REGISTER_LIMIT);
         let sp_offset: i32 = self.0.stack_pos(src_mem).try_into().unwrap();
         dynasm!(self.0.ops
             // Pretend that we're a double
@@ -49,7 +49,7 @@ impl Assembler for IntervalAssembler {
         );
     }
     fn build_store(&mut self, dst_mem: u32, src_reg: u8) {
-        assert!(src_reg < REGISTER_LIMIT);
+        assert!((src_reg as usize) < REGISTER_LIMIT);
         let sp_offset: i32 = self.0.stack_pos(dst_mem).try_into().unwrap();
         dynasm!(self.0.ops
             // Pretend that we're a double
